pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["sergei <sergeypaly4@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
django-environ = "^0.11.2"
loguru = "^0.7.2"
gunicorn = "^21.2.0"
psycopg2 = "^2.9.9"
aiogram = "^3.4.1"
redis = "^5.0.2"
pillow = "^10.3.0"
pytz = "^2024.1"
apscheduler = "^3.10.4"
apscheduler-di = "^0.1.0"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.3.0"}
babel = "^2.15.0"
django-liveconfigs = "^1.0.5"
django-redis = "^5.4.0"
django-extensions = "^3.2.3"
djangorestframework = "^3.15.2"
djoser = "^2.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
